WIP: Major refactoring to make Async::Enumerable includable in user classes
This commit represents a significant architectural refactoring of the async_enumerable gem
to support two usage patterns:
1. Direct usage on arrays/enumerables: [1,2,3].async.map { ... }
2. Includable module for custom classes: class MyCollection; include Async::Enumerable; end

## Key Changes

### Module Reorganization
- Created `Async::Enumerable::Methods` as a container module for all async method implementations
- Organized methods into logical groups following the naming pattern of actors (Transformers,
  Aggregators, etc.) with the exception of Predicates (Predicators would be too uncommon):
  - `Methods::Transformers`: map, select, reject, filter_map, flat_map, compact, uniq, sort, sort_by
  - `Methods::Predicates`: all?, any?, none?, one?, include?, find, find_index (formerly EarlyTerminable)
  - `Methods::Converters`: to_a, sync
  - `Methods::Aggregators`: (placeholder for future: reduce, sum, count, etc.)
  - `Methods::Iterators`: (placeholder for future: each_with_index, each_with_object, etc.)
  - `Methods::Slicers`: (placeholder for future: drop, grep, partition, etc.)

### Renamed Modules
- `BoundedConcurrency` → `FiberLimiter` (better describes its purpose)
- `EarlyTerminable` → `Predicates` (more descriptive of the methods it contains)
- Positioned `FiberLimiter` at the same level as `Methods` since it's a cross-cutting concern
  for the entire Enumerable module, not just the methods

### New def_enumerator Mechanism
- Added `def_enumerator` class method to allow classes to specify their enumerable source
- When a class includes `Async::Enumerable`, it can use `def_enumerator :items` to specify
  which method/attribute contains the enumerable to be made async
- Supports optional max_fibers configuration per class
- Example usage:
  ```ruby
  class TodoList
    include Async::Enumerable
    def_enumerator :todos

    def initialize
      @todos = []
    end
    attr_reader :todos
  end

  list = TodoList.new
  list.async.map { |todo| process(todo) }.sync
  ```

### Async::Enumerator Changes
- Removed direct method implementations (now in modules)
- Now includes `Enumerable::Methods` and `Enumerable::FiberLimiter`
- Kept core functionality: initialization, each implementation, comparison operators
- Methods are now modular and can be selectively included if needed

### File Structure Changes
- Each method now lives in its own module file for better organization
- Directory structure mirrors the module hierarchy:
  ```
  lib/async/enumerable/
  ├── fiber_limiter.rb
  ├── methods.rb
  └── methods/
      ├── transformers.rb
      ├── transformers/
      │   ├── map.rb
      │   ├── select.rb
      │   └── ...
      ├── predicates.rb
      ├── predicates/
      │   ├── all.rb
      │   ├── any.rb
      │   └── ...
      └── ...
  ```

### Test Structure Updates
- Moved tests to mirror the new lib structure
- `spec/async/enumerable/early_terminable/` → `spec/async/enumerable/methods/predicates/`
- `spec/async/enumerable/max_fibers_spec.rb` → `spec/async/enumerable/fiber_limiter_spec.rb`
- `spec/async/enumerable/async_spec.rb` → `spec/async/enumerator_spec.rb`

### Benefits of This Refactoring
1. **Modularity**: Methods are organized by purpose and can be selectively included
2. **Includable**: Classes can now include Async::Enumerable to add async capabilities
3. **Maintainability**: Each method in its own file makes the codebase easier to navigate
4. **Extensibility**: New method categories can be easily added (Aggregators, Iterators, etc.)
5. **Backwards Compatible**: All existing usage patterns still work

### Implementation Notes
- All 160 tests pass without modification (only moved to new locations)
- The `take_while` method was removed from EarlyTerminable as it was already delegated
  via Forwardable and doesn't benefit from async execution
- Created `scripts/` directory for test scripts that don't need to be maintained
- Fixed method chaining to ensure transformation methods return Async::Enumerator instances

### TODO/Future Work
- Implement actual methods for Aggregators, Iterators, and Slicers modules
- Add tests specifically for the includable module pattern
- Consider whether some sequential methods should be in their own module
- Document the new module structure in README
