+++ lib/async/enumerable/methods/aggregators.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Async
+  module Enumerable
+    module Methods
+      # Aggregators contains async implementations of enumerable aggregation methods
+      # that reduce collections to single values.
+      module Aggregators
+        # TODO: Add aggregation methods like reduce, sum, count, etc.
+      end
+    end
+  end
+end
