+++ lib/async/enumerable/methods/iterators.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Async
+  module Enumerable
+    module Methods
+      # Iterators contains async implementations of enumerable iteration helper methods.
+      module Iterators
+        # TODO: Add iteration methods like each_with_index, each_with_object, etc.
+      end
+    end
+  end
+end
