+++ lib/async/enumerable/methods/slicers.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Async
+  module Enumerable
+    module Methods
+      # Slicers contains async implementations of enumerable slicing and filtering methods.
+      module Slicers
+        # TODO: Add slicing methods like drop, grep, partition, etc.
+      end
+    end
+  end
+end
