+++ spec/async/enumerable/methods/predicates/all_spec.rb
