+++ spec/async/enumerable/methods/predicates/any_spec.rb
