+++ spec/async/enumerable/methods/predicates/find_index_spec.rb
