+++ spec/async/enumerable/methods/predicates/find_spec.rb
