+++ spec/async/enumerable/methods/predicates/none_spec.rb
