+++ spec/async/enumerable/methods/predicates/one_spec.rb
